backend/requirements.txt
pandas>=2.0
numpy>=1.24
emoji>=2.0
flask>=3.0
celery[redis]>=5.3
redis>=5.0
sqlalchemy>=2.0
psycopg[binary]>=3.1
flask-sqlalchemy>=3.1
boto3>=1.34
flask-cors>=4.0
flask-limiter[redis]>=3.0
python-magic>=0.4
python-dotenv>=1.0
openai>=1.0
